Allow for the node id not being set, see https://github.com/emoncms/emoncms/issues/155

+++ Modules/input/input_controller.php
@@ -208,7 +208,7 @@ function input_controller()
         {
             $valid = true; $error = "";
 
-            $nodeid = get('node');
+            $nodeid = (int) get('node');
 
             $error = " old".$max_node_id_limit;
